Add a note about using the DirectoryObject APIs
There are cases where a developer wanting to support groups in an
application doesn't intend to make a distinction between supporting
user accounts and service principals, but may accidentally introduce
a difference simply because the 'user' APIs are significantly further
up the search results, and don't reference the more general API which
works for both kinds of identities. Here I add a note and links to the
directoryObject APIs for two such pages.